requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Code Project\Decrypt_AES\aestool.py: 4,5,6
PyQt5_sip == 12.9.0

# D:\Code Project\Decrypt_AES\aestool.py: 10,11
# D:\Code Project\Decrypt_AES\libs\aes.py: 3,4
# D:\Code Project\Decrypt_AES\testAES.py: 3,4
pycryptodomex == 3.11.0

# D:\Code Project\Decrypt_AES\decryptcli.py: 3
pyperclip == 1.8.2

# D:\Code Project\Decrypt_AES\aestool.py: 13
# D:\Code Project\Decrypt_AES\testAES.py: 6
qrcode == 7.3.1
